// Source/Main/src/DMS/src/app/xoonit-share/processing-form/consts/material-control-type.enum.ts
export enum MaterialControlType {
    INPUT = 'input',
    DATEPICKER = 'datepicker',
    AUTOCOMPLETE = 'autocomplete',
    SLIDE_TOGGLE = 'slide-toggle',
    SELECT = 'select',
    RADIO_BUTTON = 'radio-button',
    IMAGE_SIGNATURE = 'signature',
    HTML_INPUT = 'html',
    SELECT_SEARCH = 'comboboxfilter',
    CHECKBOX = 'checkbox',
}